Add {data.table} to example benchmarks for comparison
It would be very cool to include {data.table} into the comparison since it is really performant and adds only one dependency. For a fair comparison the author would have to run the benchmark on the own machine, please. {data.table} is very common, as common as {dplyr}, with a very nice syntax.